feat: define wallet feture related entity

+++ src/main/java/com/ugts/transaction/service/impl/TransactionServiceImpl.java
@@ -68,7 +68,7 @@ public class TransactionServiceImpl implements TransactionService {
                 .transNo(id)
                 .bankCode(transactionRequest.getBankCode())
                 .cardType(transactionRequest.getCardType())
-                .amount((int) order.getPost().getProduct().getPrice())
+                .amount(order.getPost().getProduct().getPrice())
                 .currency(transactionRequest.getCurrency())
                 .bankAccountNo(transactionRequest.getBankAccountNo())
                 .bankAccount(transactionRequest.getBankAccount())
